requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.3
python-dotenv==1.0.0
google-generativeai>=0.3.0
requests==2.31.0
aiohttp>=3.8.0
pydantic>=2.0
